+++ src/main/java/theflogat/technomancy/common/tiles/dynamos/TileFlowerDynamo.java
@@ -18,7 +18,7 @@ public class TileFlowerDynamo extends TileDynamoBase implements IManaReceiver {
 	@Override
 	public int extractFuel(int ener) {
 		if (mana == 0) return 0;
-		float ratio = ((float) ener) / 80F;
+		float ratio = (ener) / 80F;
 		int val = (int) (20 * ratio);
 		float fuel = (float) val / (float) Math.min(mana, val);mana -= Math.min(mana, val);
 		return (int) (((float)10*20)*fuel);
